Broaden search result for other sorting than "relevance"
This is relevant for search queries containing two words or more where
the words are found in different product attributes. In this case, there
is a mechanism in the search result which searches for the single words
if no results are found otherwise. The mechanism was only used for search
by relevance which is extended to other sort orders with this commit.
Additionally, there was a bug in this mechanism - the parameter "mm=0%"
has to be set for these additional requests.